requirements.txt
matplotlib
pillow
pywinstyles
sv_ttk
python-dotenv
aiosqlite
